refactor: altera estrutura de pastas
