+++ package.json
@@ -12,7 +12,9 @@
     "lint:fix": "eslint \"src/**\" --fix"
   },
   "lint-staged": {
-    "*.tsx": ["yarn lint:fix"]
+    "*.tsx": [
+      "yarn lint:fix"
+    ]
   },
   "dependencies": {
     "@emotion/react": "^11.10.0",
@@ -32,8 +34,8 @@
     "@vitejs/plugin-react": "^2.0.0",
     "eslint": "^7.32.0 || ^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-import-resolver-typescript": "^3.4.0",
     "eslint-config-prettier": "^8.5.0",
+    "eslint-import-resolver-typescript": "^3.4.0",
     "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.2.1",
@@ -43,6 +45,7 @@
     "lint-staged": "^13.0.3",
     "prettier": "2.7.1",
     "typescript": "^4.6.4",
-    "vite": "^3.0.0"
+    "vite": "^3.0.0",
+    "vite-tsconfig-paths": "^3.5.0"
   }
 }
